Substring and RegEx Functions in SQL Server
Databases have the data stored in an organized manner. However, we may come across situations where we need to retrieve the data but we don’t have a specific filter to achieve this. For such cases, Substring and RegEx (Regular Expressions) functions can help us to retrieve the data we need. Substring Functions returns a portion of a character string, while RegEx are generalized expressions used to match patterns with various sequences of characters. See a few examples below.